Fix codacy and some header issues.
